+++ src/video_core/renderer_opengl/texture_filters/nearest_neighbor/nearest_neighbor.h
@@ -14,7 +14,7 @@ class NearestNeighbor : public TextureFilterBase {
 public:
     static constexpr std::string_view NAME = "Nearest Neighbor";
 
-    explicit NearestNeighbor(u16 scale_factor);
+    explicit NearestNeighbor(u32 scale_factor);
     void Filter(const OGLTexture& src_tex, Common::Rectangle<u32> src_rect,
                 const OGLTexture& dst_tex, Common::Rectangle<u32> dst_rect) override;
