+++ src/video_core/renderer_opengl/texture_filters/scale_force/scale_force.h
@@ -14,7 +14,7 @@ class ScaleForce : public TextureFilterBase {
 public:
     static constexpr std::string_view NAME = "ScaleForce";
 
-    explicit ScaleForce(u16 scale_factor);
+    explicit ScaleForce(u32 scale_factor);
     void Filter(const OGLTexture& src_tex, Common::Rectangle<u32> src_rect,
                 const OGLTexture& dst_tex, Common::Rectangle<u32> dst_rect) override;
