+++ src/video_core/renderer_opengl/texture_filters/xbrz/xbrz_freescale.h
@@ -13,7 +13,7 @@ class XbrzFreescale : public TextureFilterBase {
 public:
     static constexpr std::string_view NAME = "xBRZ freescale";
 
-    explicit XbrzFreescale(u16 scale_factor);
+    explicit XbrzFreescale(u32 scale_factor);
     void Filter(const OGLTexture& src_tex, Common::Rectangle<u32> src_rect,
                 const OGLTexture& dst_tex, Common::Rectangle<u32> dst_rect) override;
